refactor: rename old_codegen -> codegen (#2595)
Also rename `codegen.parser_utils` to `codegen.core`
