+++ tests/compiler/test_sha3_32.py
@@ -1,5 +1,5 @@
+from vyper.codegen.lll_node import LLLnode
 from vyper.lll import compile_lll, optimizer
-from vyper.old_codegen.parser_utils import LLLnode
 
 
 def test_sha3_32():
