+++ tests/parser/types/test_identifier_naming.py
@@ -1,8 +1,8 @@
 import pytest
 
 from vyper.builtin_functions import BUILTIN_FUNCTIONS
+from vyper.codegen.expr import BUILTIN_CONSTANTS, ENVIRONMENT_VARIABLES
 from vyper.exceptions import NamespaceCollision, StructureException, SyntaxException
-from vyper.old_codegen.expr import BUILTIN_CONSTANTS, ENVIRONMENT_VARIABLES
 from vyper.semantics.namespace import RESERVED_KEYWORDS
 from vyper.utils import FUNCTION_WHITELIST
