+++ vyper/builtin_functions/convert.py
@@ -4,9 +4,7 @@ from decimal import Decimal
 
 from vyper import ast as vy_ast
 from vyper.builtin_functions.signatures import signature
-from vyper.evm.opcodes import version_check
-from vyper.exceptions import InvalidLiteral, StructureException, TypeMismatch
-from vyper.old_codegen.parser_utils import (
+from vyper.codegen.core import (
     LLLnode,
     add_ofst,
     clamp_basetype,
@@ -16,7 +14,9 @@ from vyper.old_codegen.parser_utils import (
     load_op,
     shr,
 )
-from vyper.old_codegen.types import BaseType, ByteArrayType, StringType, get_type
+from vyper.codegen.types import BaseType, ByteArrayType, StringType, get_type
+from vyper.evm.opcodes import version_check
+from vyper.exceptions import InvalidLiteral, StructureException, TypeMismatch
 from vyper.utils import DECIMAL_DIVISOR, MemoryPositions, SizeLimits
