+++ vyper/cli/vyper_lll.py
@@ -3,9 +3,9 @@ import argparse
 import sys
 
 import vyper
+from vyper.codegen.lll_node import LLLnode
 from vyper.lll import compile_lll, optimizer
 from vyper.lll.s_expressions import parse_s_exp
-from vyper.old_codegen.parser_utils import LLLnode
 
 
 def _parse_cli_args():
