+++ vyper/codegen/expr.py
@@ -2,18 +2,8 @@ import math
 from decimal import Decimal
 
 from vyper import ast as vy_ast
-from vyper.evm.opcodes import version_check
-from vyper.exceptions import (
-    CompilerPanic,
-    EvmVersionException,
-    StructureException,
-    TypeCheckFailure,
-    TypeMismatch,
-)
-from vyper.old_codegen import external_call, self_call
-from vyper.old_codegen.keccak256_helper import keccak256_helper
-from vyper.old_codegen.lll_node import LLLnode
-from vyper.old_codegen.parser_utils import (
+from vyper.codegen import external_call, self_call
+from vyper.codegen.core import (
     clamp_basetype,
     get_element_ptr,
     get_number_as_fraction,
@@ -22,7 +12,9 @@ from vyper.old_codegen.parser_utils import (
     make_setter,
     unwrap_location,
 )
-from vyper.old_codegen.types import (
+from vyper.codegen.keccak256_helper import keccak256_helper
+from vyper.codegen.lll_node import LLLnode
+from vyper.codegen.types import (
     ArrayLike,
     BaseType,
     ByteArrayLike,
@@ -36,6 +28,14 @@ from vyper.old_codegen.types import (
     TupleType,
     is_base_type,
     is_numeric_type,
+)
+from vyper.evm.opcodes import version_check
+from vyper.exceptions import (
+    CompilerPanic,
+    EvmVersionException,
+    StructureException,
+    TypeCheckFailure,
+    TypeMismatch,
 )
 from vyper.semantics.types import DynamicArrayDefinition
 from vyper.utils import (
