+++ vyper/codegen/function_definitions/utils.py
