+++ vyper/codegen/global_context.py
@@ -2,8 +2,8 @@ from typing import Optional
 
 from vyper import ast as vy_ast
 from vyper.ast.signatures.function_signature import ContractRecord, VariableRecord
+from vyper.codegen.types import InterfaceType, parse_type
 from vyper.exceptions import CompilerPanic, InvalidType, StructureException
-from vyper.old_codegen.types import InterfaceType, parse_type
 from vyper.typing import InterfaceImports
 from vyper.utils import cached_property
