+++ vyper/codegen/lll_node.py
@@ -2,10 +2,10 @@ import re
 from enum import Enum, auto
 from typing import Any, List, Optional, Tuple, Union
 
+from vyper.codegen.types import BaseType, NodeType, ceil32
 from vyper.compiler.settings import VYPER_COLOR_OUTPUT
 from vyper.evm.opcodes import get_comb_opcodes
 from vyper.exceptions import CompilerPanic
-from vyper.old_codegen.types import BaseType, NodeType, ceil32
 from vyper.utils import VALID_LLL_MACROS, cached_property
 
 # Set default string representation for ints in LLL output.
