+++ vyper/codegen/parser.py
@@ -2,19 +2,19 @@ from typing import Any, List, Optional, Tuple, Union
 
 from vyper import ast as vy_ast
 from vyper.ast.signatures.function_signature import FunctionSignature
+from vyper.codegen.core import make_setter
+from vyper.codegen.function_definitions import (
+    generate_lll_for_function,
+    is_default_func,
+    is_initializer,
+)
+from vyper.codegen.global_context import GlobalContext
+from vyper.codegen.lll_node import LLLnode
 from vyper.exceptions import (
     EventDeclarationException,
     FunctionDeclarationException,
     StructureException,
 )
-from vyper.old_codegen.function_definitions import (
-    generate_lll_for_function,
-    is_default_func,
-    is_initializer,
-)
-from vyper.old_codegen.global_context import GlobalContext
-from vyper.old_codegen.lll_node import LLLnode
-from vyper.old_codegen.parser_utils import make_setter
 from vyper.semantics.types.function import FunctionVisibility, StateMutability
 from vyper.typing import InterfaceImports
 from vyper.utils import LOADED_LIMITS
